do not import obspy in runtests script
